pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]

[project]
name = "ChimeraX-copick"
version = "0.1.0"
dynamic = ["classifiers", "requires-python"]
dependencies = [
    "ChimeraX-Core~=1.7.1",
    "ChimeraX-ArtiaX",
    "ChimeraX-OME-Zarr",
    "pydantic",
    "hatchling",
    "copick @ git+https://github.com/uermel/copick.git",
]
authors = [
  {name = "Utz H. Ermel", email = "utz@ermel.me"},
]
description = "ChimeraX extension for collaborative annotation of cryo-electron tomograms."
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["cryoet", "cryo-et", "tomography", "annotation", "segmentation", "collaborative", "copick"]

[project.urls]
Repository = "https://github.com/uermel/chimerax-copick.git"
Issues = "https://github.com/uermel/chimerax-copick/issues"

[project.optional-dependencies]
dev = [
    "black",
    "ipython",
    "pre-commit",
    "ruff",
]

[chimerax]
category = "General"
package = "chimerax.copick"
min-session-version="1"
max-session-version="1"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
]

[chimerax.tool."Copick"]
tool_name = "Copick"
category = "General"
synopsis = "Collaborative annotation of cryo-electron tomograms."

[chimerax.command."copick start"]
category = "General"
synopsis = "Start a copick session or load a new config file."

[tool.black]
line-length = 120
target_version = ['py311']

[tool.ruff]
select = [
    "E", "W",  # pycodestyle
    "F",  # pyflakes
    "B",  # bugbear
    "I",  # isort
    "N",  # pep8-naming
    "ASYNC",  # async
    "COM",  # commas
    "C4",  # comprehensions
    "DTZ",  # datetimez
    "ISC",  # implicit-str-concat
    "G",  # logging-format
    "T20",  # print
    "SIM",  # simplify
]
ignore = [
    "N802",  # Function name should be lowercase
    "N806",  # Variable in function should be lowercase
    "B008",  # Do not perform calls in argument defaults
    "N999",  # Invalid module name
    "E501",  # line too long
    "C408",  # rewrite empty built-ins as literals
    "T201",  # print statements.
    "DTZ007", # Datetime objects without timezones.
    "DTZ005", # More datetimes without timezones.
]
line-length = 120
target-version = "py311"

[tool.ruff.isort]
known-first-party =["common"]

[tool.ruff.flake8-quotes]
docstring-quotes = "double"

[tool.mypy]
show_error_codes = true
ignore_missing_imports = true
warn_unreachable = true
strict = true
